+++ example/lib/line_chart/samples/line_chart_sample6.dart
@@ -90,7 +90,7 @@ class LineChartSample6 extends StatelessWidget {
                     ),
                     dotData: FlDotData(
                       show: true,
-                      getDotColor: (_) => Colors.deepOrange.withOpacity(0.5),
+                      getDotColor: (spot, percent, barData) => Colors.deepOrange.withOpacity(0.5),
                       dotSize: 12,
                     ),
                   ),
@@ -108,7 +108,7 @@ class LineChartSample6 extends StatelessWidget {
                     ),
                     dotData: FlDotData(
                       show: true,
-                      getDotColor: (_) => Colors.blue.withOpacity(0.5),
+                      getDotColor: (spot, percent, barData) => Colors.blue.withOpacity(0.5),
                       dotSize: 12,
                     ),
                   ),