+++ lib/src/utils/utils.dart
@@ -76,3 +76,28 @@ BorderRadius normalizeBorderRadius(BorderRadius borderRadius, double width) {
     bottomRight: bottomRight,
   );
 }
+
+/// Lerps between a [LinearGradient] colors, based on [t]
+Color lerpGradient(List<Color> colors, List<double> stops, double t) {
+
+  if (stops == null || stops.length != colors.length) {
+    stops = [];
+    /// provided gradientColorStops is invalid and we calculate it here
+    colors.asMap().forEach((index, color) {
+      final percent = 1.0 / colors.length;
+      stops.add(percent * (index + 1));
+    });
+  }
+
+  for (var s = 0; s < stops.length - 1; s++) {
+    final leftStop = stops[s], rightStop = stops[s + 1];
+    final leftColor = colors[s], rightColor = colors[s + 1];
+    if (t <= leftStop) {
+      return leftColor;
+    } else if (t < rightStop) {
+      final sectionT = (t - leftStop) / (rightStop - leftStop);
+      return Color.lerp(leftColor, rightColor, sectionT);
+    }
+  }
+  return colors.last;
+}