+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fixed a minor touch bug in the [BarChart](https://github.com/imaNNeoFighT/fl_chart/blob/master/repo_files/documentations/bar_chart.md).
 * Fixed ScatterChart built-in touch behaviour.
 * Fixed drawing grid lines bug, #280.
+* Implemented [FlDotData](https://github.com/imaNNeoFighT/fl_chart/blob/master/repo_files/documentations/line_chart.md#fldotdata).`getDotColor` in a proper way, it returns a color based on the [LineChartBarData](https://github.com/imaNNeoFighT/fl_chart/blob/master/repo_files/documentations/line_chart.md#linechartbardata) color, #274, #282.
 
 ## 0.8.7
 * Added `show` property in the `VerticalLineLabel` and set default to `false`, #256.